+++ extension/src/extension.ts
@@ -1,3 +1,4 @@
+import cacache = require('cacache');
 import * as vscode from 'vscode';
 import * as nls from 'vscode-nls';
 
@@ -8,11 +9,8 @@ import { ExtensionInfoService } from './extensionInfo';
 import { ExtensionsFileFeatures } from './extensionsFileFeatures';
 import { RegistryProvider } from './RegistryProvider';
 import { UpdateChecker } from './UpdateChecker';
-import { deleteNpmDownloads } from './util';
+import { deleteNpmDownloads, getNpmCacheDir } from './util';
 import { RegistryView } from './views/registryView';
-
-// TODO: provide autocomplete suggestions in extensions.private.json for
-// registries: { "name": "Registry Name", "registry": "https://my-private.registry" }
 
 // TODO: notify user if extensions.private.json recommends extensions that are
 // not installed. Add a way to ignore for a workspace or disable globally.
@@ -38,6 +36,8 @@ export function activate(context: vscode.ExtensionContext) {
         registerCommands(registryProvider, registryView, updateChecker, extensionInfo),
         registerLanguageFeatures(registryProvider),
     );
+
+    verifyCache();
 }
 
 export async function deactivate() {
@@ -85,4 +85,12 @@ function registerCommands(
 
 function registerLanguageFeatures(registryProvider: RegistryProvider): vscode.Disposable {
     return vscode.Disposable.from(new ExtensionsFileFeatures(registryProvider));
+}
+
+async function verifyCache() {
+    const cache = getNpmCacheDir();
+    if (cache) {
+        const stats = await cacache.verify(cache);
+        console.log('Cleaned private extension manager cache:', stats);
+    }
 }